test: update tests with folder refactor
