+++ Website/tests/components/GridView/GridView.test.tsx
@@ -3,7 +3,7 @@ import { render, fireEvent  } from "@testing-library/react";
 import { expect, it } from "vitest";
 import { HashRouter as Router } from 'react-router-dom';
 
-import GridView from '../../../src/components/GridView';
+import GridView from '../../../src/components/Views/Grid/GridView';
 import mockData from '../../../src/mockData'
 
 it('GridView renders correctly', () => {
