+++ Website/tests/components/Homepage/Homepage.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from "@testing-library/react";
 import { expect, it } from "vitest";
 
-import Homepage from '../../../src/components/Homepage';
+import Homepage from '../../../src/components/Homepage/Homepage';
 
 it('Homepage renders correctly', () => {
     const tree = render(<Homepage/>)
