+++ Website/tests/components/ListView/ListView.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from "@testing-library/react";
 import { expect, it } from "vitest";
 
-import ListView from '../../../src/components/ListView';
+import ListView from '../../../src/components/Views/List/ListView';
 import mockData from '../../../src/mockData'
 
 it('ListView renders correctly', () => {
